+++ src/chart/HoldNoteObject.cs
@@ -7,7 +7,7 @@ public partial class HoldNoteObject : NoteObject {
     public required NoteTime EndTime;
     public float PixelsPerBeat;
     public float JudgementY;
-    public int BeatsPerMeasure;
+    public long BeatsPerMeasure;
     private TextureRect? _trail;
     private Sprite2D? _noteSprite;
     private bool _held;
@@ -30,13 +30,13 @@ public partial class HoldNoteObject : NoteObject {
         }
     }
 
-    public void StartHold(NoteType type, int bar, int beat, double sixteenth) {
+    public void StartHold(NoteType type, long bar, long beat, double sixteenth) {
         if (IsEventMatching(type, bar, beat, sixteenth)) {
             _held = true;
         }
     }
 
-    public void EndHold(NoteType type, int bar, int beat, double sixteenth) {
+    public void EndHold(NoteType type, long bar, long beat, double sixteenth) {
         if (IsEventMatching(type, bar, beat, sixteenth)) {
             _held = false;
         }
